CRUD operations

+++ src/main/java/com/nous/app/repo/CalendarRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.nous.app.model.Holiday;
 
-public interface CalendarRepository extends CrudRepository<Holiday, Integer>{
+public interface CalendarRepository extends CrudRepository<Holiday, Long>{
 
 	public List<Holiday> findByName(String name);
